#13 Refactor project model and mapping functions; consolidate entity conversion logic
