+++ backend/src/main/kotlin/dev/ecckea/agilepath/backend/domain/project/model/NewProject.kt
@@ -1,7 +1,5 @@
 package dev.ecckea.agilepath.backend.domain.project.model
 
-import dev.ecckea.agilepath.backend.domain.project.repository.entity.ProjectEntity
-import dev.ecckea.agilepath.backend.domain.user.repository.entity.UserEntity
 import dev.ecckea.agilepath.backend.shared.utils.now
 import java.time.Instant
 
@@ -11,12 +9,4 @@ data class NewProject(
     val framework: Framework,
     val createdBy: String, // UserId
     val createdAt: Instant = now(),
-)
-
-fun NewProject.toEntity(createdBy: UserEntity): ProjectEntity = ProjectEntity(
-    name = name,
-    description = description,
-    framework = framework,
-    createdBy = createdBy,
-    createdAt = createdAt
 )
